refactor: move tests into pagination
